winget: auto-update license/copyright urls
One of our winget manifests currently has hardcoded, outdated values for
its LicenseUrl and CopyrightUrl fields. This is because the winget-create
tool that we use to automatically generate and publish our manifests
cannot change these fields.

In light of this, instead of creating and submitting manifests with a
single command, update the workflow to do the following:

1. Create the manifests.
2. Update the LicenseUrl and CopyrightUrl fields with the correct version.
3. Submit the manifests.

This change also adds the workflow_dispatch trigger to the release-winget
workflow for easier testing in the future.

Signed-off-by: Lessley Dennington <ledennin@microsoft.com>