release: remove pmc deployment workflow
The Debian package for each microsoft/git release is currently deployed
to packages.microsoft.com. The team that manages this feed is migrating
to a new "vNext Publishing API" [1]. Because the new API has several
administrative pre-requisites and the maintainers believe consumption of
microsoft/git from this feed to be negligible, the decision was made to
cease deploying to packages.microsoft.com rather than migrate to the new
API. This change removes the workflow that handled this deployment.

1: https://microsoft.sharepoint.com/teams/PMC/SitePages/API-An.aspx