gvfs-helper: show progress of multiple prefetch packs
When using a cache server to download multiple prefetch packs, the GVFS
Protocol sends all packs in one download. This download is split into
multiple files during that single HTTPS request. This also presents a
single progress indicator such as

  Prefetch 1691384600 (2023-08-07 05:03:20 +0000) (bytes received): 25355702, done.

After downloading all of these packs to the `tempPacks` directory of the
shared object cache, we run `git index-pack` on each packfile
individually.

The previous change removed the verbose output of `git index-pack`, but
that left the time spend indexing the packs without any progress
indicator.

Add a new progress indicator that ticks through the different prefetch
pack-files that are being installed as we go. This presents a new
progress indicator that terminates with output like

  Installing prefetch packfiles: 100% (14/14), done.

This helps users understand what is going on without the extra noise of
two output lines per `git index-pack` command.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>