fixup! scalar: do initialize `gvfs.sharedCache`
In this commit, we added the 'credential.interactive=never' option to
unattended scalar options. This should be changed to 'false' to match
the modern use of this config option.

But also, we have a test that requires using askpass to get credentials,
but the test is in unattended mode. Fix that test to include
'credential.interactive=true' to bypass this issue.