index-info.c: return unrecognized lines to caller
Update 'read_index_info()' to return INDEX_INFO_UNRECOGNIZED_LINE (value 1),
rather than die()-ing when the function encounters a line that cannot be
parsed according to one of the accepted formats. This grants the caller the
flexibility to fall back on custom handling for such lines rather than a
returning a catch-all error. In the case of 'update-index', we'll still exit
with a "malformed input line" error. However, when 'read_index_info()' is
used to process the input to 'mktree' in a later patch, an empty line return
value will signal a new tree in --batch mode.

Signed-off-by: Victoria Dye <vdye@github.com>